Restore the all-links crawl snapshot (verifySteps)
The exact-order step list was removed in a63b840 because it was flaky —
but the flake had a cause: the crawler collects a page's links immediately,
while kolay compiles markdown asynchronously, so in-content links were
discovered (or not) depending on timing. The original 250ms settle delay
was what made discovery deterministic; restoring it makes the 42-step
snapshot stable (8/8 consecutive green runs).

Doing so surfaced a second, pre-existing issue the old single-test-file
layout could never show: fully compiling many pages and tearing the app
down mid-flight wedges ember-repl's module-level compiler, so a subsequent
test app's runtime-compiled pages never render (even with a 10s waitFor).
Until that's fixed upstream, the crawl file is named zz-* so it runs last.
